move Sequence in a helper function
